Adding divisor_override and ceil_mode functionality to avg_pool2d (#20513)
### Ticket
[16740](https://github.com/tenstorrent/tt-metal/issues/16740)

### Problem description
Pytorch
[implementation](https://pytorch.org/docs/stable/generated/torch.nn.AvgPool2d.html)
of the avg_pool2d requires divisor_override as an optional int argument
to be used instead of kernel_size when calculating the average. Ttnn
implementation was missing this argument. The ceil_mode set to true did
not have supported functionality for avg_pool operation. More than one
scalar need to be filled for avg pool if the ceil_mode is true.

### What's changed
This PR is introducing breaking change for max_pool2d and avg_pool2d
users. Besides adding a new argument it splits the avg and max pool
calls in python altogether.
Changes to avg_pool2d
- Multiple scalars are sent to each core as runtime arguments depending
on the output sticks cores need to produce as a support of the
ceil_mode=true instead of one.
- Dilation is removed as it is not needed
- Divisor_override is added as a new optional int argument with the full
functionality supported
- Reorder of the arguments to match the signature of the pytorch,
ceil_mode is moved from the end of the argument list to right after
padding
Changes introduced to max_pool2d
- Reorder of the arguments to match the signature of the pytorch,
ceil_mode is moved from the end of the argument list to right after
dilation

### Additional
Unused argument from construct_on_host_config_tensor in sliding_window
is removed

### Implementation details
Since ceil mode set to true argument needs more than one scalar for the
result to be valid due to the fact that we must exclude the padding
coming from ceil mode when dividing with kernel_size, dividing by only
the pool area covering valid tensor datums and none of the padding,
there was a need of calculating the array of scalars and times when they
need to be fetched for each core and passed as a config tensor to reader
kernels. Both reader kernels are doing the filling of the scalar value
for the scalars needed for even or odd sticks calculation each in its
own scalar cb. Readers are getting array with elements consisting of 3
pieces of information: scalar start, scalar value and scalar end. Start
and end are integer determining interval on which scalar is valid for
the nsticks to be calculated **[start, end)** whereas compute kernel is
waiting on each element. The filling of the correct value is happening
only when the value changes so in cases when the value is the same
reader does only reserve, push back. This is why only one page is used
for the scalar cbs. The "time when the scalar needs to be changed" as
well as the scalar itself is calculated on the host side and packed to
config tensor.

For the maxpool and avg pool ops that need only one scalar different
path is used not to take up L1. These paths are divided by constexpr
compile time argument so the ifs will be removed conditionally during
the compilation of the kernels.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/15464394731)
- [x] [Nightly tt-metal L2 tests
WH](https://github.com/tenstorrent/tt-metal/actions/runs/15464404808)
- [x] [Nightly tt-metal L2 tests
BH](https://github.com/tenstorrent/tt-metal/actions/runs/15319217876)
- [x] [(Single-card) Device perf
regressions](https://github.com/tenstorrent/tt-metal/actions/runs/15462630989)
- [x] [(Single-card) Frequent model and ttnn
tests](https://github.com/tenstorrent/tt-metal/actions/runs/15462610399)