Don't use transpose_shards in BS conv2d
Issue #17897

On WH and BH arch dram cores are placed as columns
on the grid. To get good dram bandwidth we need
dram readers to be perpendicular to dram cores,
meaning they need to be placed as rows on the grid.

transpose_shards on Block Sharded Conv2d was causing
dram readers to be placed as columns on the grid.

Change default for transpose_shards to False in
Conv2dConfig, and remove usage of transpose_shards
from model code.

Resolved reamining issues in code when transpose_shards
== False.

Primary motivation for this change is to improve
unet performance of SDXL on WH arch.
Up to 2x performance on block sharded conv2ds in
unet SDXL.

Other models that have improved device performance:
vgg, ufld_v2, vanilla_unet, yolov10