[skip ci] Enforce =delete (#23156)
### Ticket
#23155

### Problem description
To quote the clang-tidy check's description:

> Before the introduction of C++11, the primary method to effectively
“erase” a particular function involved declaring it as private without
providing a definition. This approach would result in either a compiler
error (when attempting to call a private function) or a linker error
(due to an undefined reference).
>
> However, subsequent to the advent of C++11, a more conventional
approach emerged for achieving this purpose. It involves flagging
functions as = delete and keeping them in the public section of the
class.

### What's changed
* Activated the check
* Fixed the diagnosed code