Remove tunneling reservation logic and add fabric connection args when fabric is enabled (#23080)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/18726

### Problem description
- For Fast dispatch on fabric, we connect the prefetch/dispatchers
directly as upstream and downstream kernels and tunneling reservation
logic needs to be removed so a fabric router can be created on the
active ethernet core
- Additional compile args are needed for prefetch/dispatch to connect to
fabric
- Follow up PR will be made for device changes to fully enable fabric

### What's changed
- Update Dispatch and Prefetch kernels to accept a `RelayMux` as a
downstream kernel
- Disable tunneling / active ethernet core reservation logic when FD
Fabric is enabled. This will cause a fabric router to be created on the
free active eth core
- Pass in additional compile time args required for prefetch and
dispatch to connect to the fabric mux (`RelayMux`)
- Automatically initialize the fabric config to be 1D fabric when FD
Fabric is enabled. Mixed Fabric config is not supported and will throw
an error
- The `TT_METAL_FD_FABRIC` variable will be removed when dispatch on
fabric is fully enabled

### Checklist
APC
https://github.com/tenstorrent/tt-metal/actions/runs/15499620045
BH
https://github.com/tenstorrent/tt-metal/actions/runs/15473720663
T3K
https://github.com/tenstorrent/tt-metal/actions/runs/15474208392