[skip ci] Add job to publish wheel (#23328)
### Ticket
#22986 

### Problem description
We need to publish ttnn wheel to pypi.

### What's changed
Add job in package-and-release.yaml which would publish our built wheel
to test pypi.
This is a first step to put the infrastructure in place.
After this, we just wait for a swap to pypi.

### Checklist
- [x] [Package And
Release](https://github.com/tenstorrent/tt-metal/actions/runs/15568567059)
CI passes
- [x] New/Existing tests provide coverage for changes