Check for linked transactions in noc sanitizer
Raises a noc sanitizer symptom when there a unicast is requested while
there is an open linked multicast transaction.

This feature is opt-in, controlled by an env var.

When enabled, the state of linked transaction is recorded each time
NOC_CTRL register is written, and checked each time a NOC operation
in dataflow_api.h is called.