[skip ci] Bump timeout (#23551)
### Ticket
N/A

### Problem description
Saw an instance where the compile took 23m and the overall job slammed
headfirst against the 30m timeout.
It shouldn't take that long even with 0% cache misses. But we're still
investigating degraded performance on the CPU runners (slow disk) and
the CPU is known to be oversubscribed. Better to not fail the jobs.

### What's changed
Bump the timeout.