ci: remove special case handling for private runner (#377)
### Ticket
None

### Problem description
Historically, we used a private runner that couldn't access the
Tenstorrent Harbor. To handle this, we had conditional logic in place to
avoid using Harbor with that runner. As of last week, we're no longer
using this runner, so the check can be removed and the code simplified.

### What's changed
Removed a scenario that is no longer possible.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update