fix: revert "Add instr_mod support for binary shift (#365)" (#385)
### Ticket
None

### Problem description
We've observed breaking changes in tt-metal after
https://github.com/tenstorrent/tt-llk/pull/365 got merged. See
https://github.com/tenstorrent/tt-metal/actions/runs/15742340900/job/44371023416
for more information. Since there is currently no tt-metal counterpart
to this change, we decided to revert it to unblock further tt-llk
uplifts, since we have high priority tasks to finish that are currently
stuck. Once tt-metal changes are prepared, this change can be reverted
again.
FYI @Aswinmcw 

### What's changed
This reverts commit ae8d06c7437f4024a41a331eed0d3a6cc6c2bf86.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update