#24035: Fix implementations for MeshDevice query APIs
There are a number of query APIs that verify that all devices in the
mesh contain identical values, otherwise we throw an exception.

There is a special case for N300 cards where creating two unit
MeshDevice objects can erroneously throw an exception. The problem is
that our validation logic validates over the opened devices, not the
devices owned by the MeshDevice. Hence, for N300 cards, opening a unit
MeshDevice on the non-MMIO mapped chip can result in this mismatch
between opened devices and owned devices.

The original error being reported exposes a bug when users create a unit
MeshDevice on systems where we don't have direct PCIe connection to each
chip.