[skip ci] Revert "#0: Add lsan.supp to side-step errors from mem-leaks in openmpi lib (#24025)" (#24040)
This reverts commit 37cf037f93ca18e9f5509bf6ac888ece213c7413.

This was missed b/c the logic to decide whether it's worthwhile running
tests figured no code change occurred.

The problem is that the suppressions file needs to be packaged up else
it doesn't exist in the environment that needs it. That was my
oversight. Will fix tomorrow. Tonight just un-break main.