Avoid copy init (#23555)
### Ticket
Fixes #23554

### Problem description
We're making copies of objects when we don't need to.
Most (hopefully all) are small enough that it shouldn't move the needle
on perf.

CODEOWNERS: Please comment if you see any of these copies that would be
particularly expensive; either large amounts of data, or performed in a
tight loop where small bits add up in aggregate. Unless I hear otherwise
I'll assume there's no need to check perf with these changes.

### What's changed
* Enable the ClangTidy check
* Let ClangTidy fix the diagnosed code
* Revert certain tests that are making a copy for use as a Golden Master
* I think ClangTidy would understand it better if it received a const
for use a GM and made a copy for modifications; for this PR I left it
as-is and NOLINT'd the line.
* Excluded MemoryPin from being diagnosed. That looks like it wants to
use a copy for RAII purposes. Please correct me if I misunderstood its
API.

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/15644599157)