Add local mesh binding support to control plane
These belong to the "big-mesh" changes to be ported back into main. This
provides the supporting changes to allow processes launched with MPI to
be bound to a particular {MeshId, HostRankId} as defined in the
MeshGraphDescriptor.

This is implemented in same mechanism as MPI/torch.distributed by
configuring custom process specific environment variables that can be
made available.

These changes enable single-process to be launched with `TT_MESH_ID` and
`TT_HOST_RANK` to emulate process running on part of a larger mesh.

The next set of changes for process launcher will perform automatic MPI
rank to {MeshId, HostRankId} binding + support for multi-process
testing.