fix: do not use TensixDebug for debug register access (#415)
### Ticket
None

### Problem description
We had to do some code refactoring when we've switched to the latest
version of tt-exalens. In the process, we've changed a way how we
perform soft reset of cores inside Tensix. However, this API is gonna
get removed at some point according to
https://github.com/tenstorrent/tt-exalens/issues/422. Hence, we need to
stop depending on it.

### What's changed
Soft reset mechanism has been updated, to stop depending on
soon-to-be-removed API.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

---------

Co-authored-by: Strahinja Stanisic <sstanisic@tenstorrent.com>