Configure 6U CI pipelines to run nightly (#24281)
### Problem description
After adding specific 6U optimisations we want to monitor 6U pipelines
and not only 4U. This enables tracking perf in superset for 6U as it's
our end goal to show perf on this type of machine and also catch issues
that are 6U specific since we have more and more 6U specific
implementation

### What's changed
Galaxy frequent, model-perf pipelines changed to execute nightly instead
of weekly.
Unit and nightly (long stress test) execute on Saturday morning to relax
potential workload on working days.

Total workload of nightly pipelines should be <5hrs which is fine for
the beginning. We will monitor and if it clogs CI I will reduce
frequency for some pipelines.