Make ControlPlane Aware of Local vs. Global Distributed Scope (#24193)
### Ticket
None

### Problem description
When we launch processes with MPI, that process needs to be tied to a
specific MeshID and offset into the multi-host mesh as defined by the
MeshGraphDescriptor. We need some supporting infra to virtualize
ControlPlane across multi-host while providing useful accessors to go
between local vs. global view of the mesh.

### What's changed
These belong to the "big-mesh" changes to be ported back into main. This
provides the supporting changes to allow processes launched with MPI to
be bound to a particular {MeshId, HostRankId} as defined in the
MeshGraphDescriptor.

This is implemented in same mechanism as MPI/torch.distributed by
configuring custom process specific environment variables that can be
made available.

These changes enable single-process to be launched with `TT_MESH_ID` and
`TT_HOST_RANK` to emulate process running on part of a larger mesh.

The next set of changes for process launcher will perform automatic MPI
rank to {MeshId, HostRankId} binding + support for multi-process
testing.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
https://github.com/tenstorrent/tt-metal/actions/runs/15965699270
- [x] T3K Unit Tests:
https://github.com/tenstorrent/tt-metal/actions/runs/15965706212,
https://github.com/tenstorrent/tt-metal/actions/runs/15966435889