Relocate the TRISC kernel running write to mailbox (#422)
Write to the mailbox that the TRISC kernel is running immediately after
detecting that tensix is complete, instead of doing it in trisc.cpp

### Ticket


### Problem description
When running multiple tests one after the other on simulator, only the
first one passes, and the following ones fail. This is because at the
end of the first test there is a 1 written to the mailbox. This value
gets read, instead of the 2 which is supposed to be written into the
mailbox by trisc.cpp. This results in the second test getting
interpreted as already finished, when it has not even started yet.

### What's changed
Instead of trisc.cpp writing the 2 into the mailbox, it is now done
inside of device.py, right after detecting that Tensix operations are
complete.

### Type of change

- [x ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

### Checklist
<!-- These are required steps and need to be run from tt-metal
repository's Actions. Use links below and replace them with your run -->
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)