ci: use harbor for faster fetching of Docker images (#439)
### Ticket
None

### Problem description
We disabled harbor last week in #416 due to service being down, to
unblock our development. Since harbor is back in service, we should
restore this setting, as it speeds up execution of our jobs.

### What's changed
Use `harbor.ci.tenstorrent.net` to fetch docker images.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update