Workaround to allow DPRINT in noc_nonblocking_api.h
There will be a circular dependency risc_common.h -> noc_nonblocking_api.h -> dprint.h
if DPRINT is needed in noc_nonblocking_api.h and it won't compile because the definition
of invalidate_l1_cache is not visible to dprint.h.  The workaround is to move the
definition up a bit.