Fix included CPP files for wheel (#24534)
Since 0.59.0rc51 (and later), the released wheel file has been broken.
Specifically, it looks like there were commits that added new file paths
that were not being included in the wheel.

This PR fixes the issue by adding the new file paths. This fix has been
verified against the lowering tests in pytorch2.0_ttnn.

Note that the public release 0.60.0 currently has this issue. I am not
sure how to update the public release, but it may be worth doing so.