Separate the _new_state into an externally called create_state (#24622)
### DeepSeek implementation API change
We separate the model state creation into a separate function to
decouple it from the prefill and decode configs and 1. allow it to use
the hf_config directly and 2. for create_run_config to be able to take
the model configs separately (and not both at the same time)

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes

Signed-off-by: Bartłomiej Cieślar <bcieslar@tenstorrent.com>