Avoid pool l1 usage assert in no-dispatch mode (#24701)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/24358

### Problem description
We have an assert in pool op that checks if calculated and allocated L1
memory is the same. In NO_DISPATCH mode, we cannot check allocator state
since no memory is allocated.

### What's changed
We check if allocated memory is zero, if so, we assume it is NO_DISPATCH
mode and avoid the assert. There is no way for op to know if it is
running in NO_DISPATCH mode, and it is still being discussed how we
should resolve this problem in a clean way, but in the meanwhile we need
to unblock Forge as they are the ones using this mode.

We already do the same for conv2d.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16119373253)
passes
- [x] Nightly tt-metal L1 passes
[wormhole](https://github.com/tenstorrent/tt-metal/actions/runs/16119377991)
[blackhole](https://github.com/tenstorrent/tt-metal/actions/runs/16119384471)
- [x] [Frequent ttnn and
models](https://github.com/tenstorrent/tt-metal/actions/runs/16119390661)
- same errors as on main