Clang Tidy Light for PR Gate to shift-left some of ClangTidy's detections.
This is NOT a full Clang Tidy scan.  Some issues will not be detected here and this will be green while the full scan in Merge Gate will light up red.  But for what it does detect we can fix it earlier in the SDLC without undue pressue on the infra.