[skip ci] Restart SSH Daemon in container when restarted (#24841)
Refactor image to use an entrypoint to wrap the given command and start
SSH to persist SSH post restarts

### Ticket
Link to Github Issue
#24529 

### Problem description
We need SSH to start if the container restarts

### What's changed
This PR adds an entrypoint for the container. It wraps anything given
via the entrypoint script, execing `$@` after ensuring SSH is started

### Checklist
- I refuse to test anything. As long as PR Gate and Merge Gate are
happy, we should be good.

---------

Co-authored-by: Ryan Hartje <ryan@ryanhartje.com>
Co-authored-by: Ryan Hartje <hartjepc@gmail.com>