Adding support for mesh workload in Inspector (#24060)
There isn't a store that tracks `MeshDevice` or `MeshWorkload`. Added
one in Inspector. It will track lifetime of `MeshDevice` and
`MeshWorkload` objects. To achieve this, small set of events is added to
those classes.

This information helps tt-triage know what is happening with tt::nn
workloads (for starters).

Adding support in Inspector to log (for now) mesh device and workload
events and parsing the in python for tt-triage. In the future one would
be able to extract this information over RPC (as any other Inspector
information).