Rework BufferDistributionSpec API for providing core groups (#24909)
### Ticket

### Problem description
We need to rework BufferDistributionSpec API for providing core groups
to make it usable for the OP developers.
Based on the feedback from @bbradelTT, we need to provide CoreRangeSets
instead of vectors of cores to be compatible with the existing APIs and
approaches.

### What's changed
Changed core groups API in BufferDistributionSpec to provide cores
within each group and all cores together as CoreRangeSets
Calculate CoreGroups and some other data on BufferDistributionSpec
creation to avoid recomputing it over and over again
Minor API renaming in BufferDistributionSpec for consistency
Added CoreRangeSet constructor from a span of CoreCoords
Updated the tests to use the new API

### Checklist
- [x] [All post commit CI
passes](https://github.com/tenstorrent/tt-metal/actions/runs/16210433524)
- [x] New/Existing tests provide coverage for changes