Fix TG Failures caused by af50228
The TG systems are "special" because they feature N150 Wormhole cards that
serve as MMIO devices which are "gateways" to the 32-chip mesh. In our
software, these N150 cards should not be exposed to the users.

This is a quick fix to avoid the footgun of unintentionally accessing a
non-user accessible mesh. The more sustainable approach is to make it
impossible for these footguns to occur. However, this is outside scope
of immediate fix to resolve TG-regression failures and intention is to
deprecate TG support.