Fix TG Failures caused by 3b743f5 (#24997)
### Ticket
Link to Github Issue

### Problem description
TG Regression Tests are failing.

### What's changed
The TG systems are "special" because they feature N150 Wormhole cards
that serve as MMIO devices which are "gateways" to the 32-chip mesh. In
our software, these N150 cards should not be exposed to the users.

This is a fix to avoid the footgun of unintentionally accessing a
non-user accessible mesh. The more sustainable approach is to make it
impossible for these footguns to occur. However, this is outside scope
of immediate fix to resolve TG-regression failures and intention is to
deprecate TG support.

While `control_plane` sees all meshes as defined by the Mesh Graph
Descriptor, only the user-accessible mesh should be used for binding
processes onto.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes:
https://github.com/tenstorrent/tt-metal/actions/runs/16259421782
- [x] T3K Unit Tests:
https://github.com/tenstorrent/tt-metal/actions/runs/16259427361
- [x] TG Unit Tests:
https://github.com/tenstorrent/tt-metal/actions/runs/16258592873
- [ ] TG Frequent:
https://github.com/tenstorrent/tt-metal/actions/runs/16259919253