enable use of two txqs in TT-Fabric on Blackhole (#24678)
With this change, both TXQ0 and TXQ1 are used by Blackhole. Sender
channel traffic (payloads and departing credits) are mapped to TXQ0 and
receiver channel traffic (departing ack/completion credits) are mapped
to TXQ1.

The implementation is sub-optimal but is due to a HW bug in Blackhole
Ethernet subsystem. The bug prevents concurrent Ethernet remote register
writes across multiple TXQs. Specifically, on the receive side, register
writes over Ethernet can only be (statically) directed to one RXQ at a
time, configured through a register write. The register writes cannot be
dynamically be directed to different receivers, based on sender or
payload contents.

As a work-around, specifically for Blackhole in two TXQ operating mode,
the receiver -> send credit scheme moves to counter based credits. These
credits are sent using using the Ethernet send packet path (i.e. not a
register write) . This approach is race-safe. The reason this approach
is race-safe is because we use unbounded counters, as opposed to credits
that indicate an increment of some kind. Since it's an unbounded
counter, where the actual value is used as is directly, then it is okay
if a value is sent multiple times. A value could be sent multiple times
if the Erisc updates the values multiple times before the ethernet
subsystem is able to dispatch the write over Ethernet.

Note that in the abstract, there is a roll-over/overflow concern for
unbounded counters. In practice, this is not an issue because of how
subtraction is implemented in hardware. If the sender channel is
comparing two unsigned counters: one that has wrapped around against one
that hasn't. From the uint counter perspective, the wrap doesn't matter.
A snippet such as the following running on the risc demonstrates it.

```
    uint32_t trailing = std::numeric_limits<uint32_t>::max() - 3;
    uint32_t leading = 4; // indicating that leading wrapped recently (current offset is 4 past wrap)

    uint32_t delta = leading - trailing;
    
    uint32_t counted_delta = 0;
    for (uint32_t ctr = trailing; ctr != leading; ctr++)  {
        counted_delta++;
    }

    assert(counted_delta == delta);
```

In addition to integration into the TT-Fabric router, this approach is
implemented in the unit test for basic bringup and validation.

Additionally, a unit test to program this two-txq mode is added to test
the programming sequence to set TXQ1 to packet mode in isolation. This
unit test also loosely models the general traffic patterns sent by
fabric routers.

To support the new behaviour, the fabric router code was updated to
start formalizing these credit receivers/senders into separate types
which are compile time selectable now.

```
                          ┌─── now a compile time configurable type                 
                          │                          ───┐                           
               Chip 0     │                             │      Chip 1               
            Fabric Router │            srcId            │   Fabric Router           
   ┌──────────────────────┼───────┐     │          ┌────┼─────────────────────────┐ 
   │                      │       │     │          │    │                         │ 
   │                      ▼       │    \│          │    ▼                         │ 
   │ ┌──────────────┐──────────┐  │   │ \          │ ┌────────┬────────────────┐  │ 
   │ │ Inbound 0    │ Credit   │  │   │  \ Credit  │ │Credit  │ Inbound 0      │  │ 
   │ │ Channel      │ Receiver │◄─┼───┤  |     ┌───┼─┤Sender  │ Channel        │  │ 
   │ └───────┬──────┘──────────┘  │   │  |◄────┘   │ └────────┴────────────────┘  │ 
 ┌─┼─────────┘                    │ ┌─┤  |         │                  ▲           │ 
 │ │ ┌──────────────┐──────────┐  │ │ │  /         │                  │           │ 
 │ │ │ Inbound 1    │ Credit   │  │ │ │ /          │                  │           │ 
 │ │ │ Channel      │ Receiver │◄─┼─┘  /           │                  │           │ 
 │ │ └───────┬──────┘──────────┘  │◄──────────────►│                  │           │ 
 │ │         │                    │     Ethernet   │                  │           │ 
 │ │         │                    │                │                  │           │ 
 │ │         │                    │                │                  │           │ 
 └─┼────────►└────────────────────┼────────────────┼──────────────────┘           │ 
   │                              │                │                              │ 
   │                              │                │                              │ 
   │                              │                │                              │ 
   │                              │                │                              │ 
   │                              │                │                              │ 
   └──────────────────────────────┘                └──────────────────────────────┘ 
```

In this change, the direction of credit send/receive (relative to
Ethernet link) still hardcoded by type but in the future they will
likely be generalized further to support various interfaces
(Ethernet-subsystem, noc).



### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI: https://github.com/tenstorrent/tt-metal/actions/runs/16209520690
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI with demo tests passes (if applicable):
https://github.com/tenstorrent/tt-metal/actions/runs/16243051049
- [x] BH LLMbox:
https://github.com/tenstorrent/tt-metal/actions/runs/16243041071
- note on reruns: Each run passed. Wanted to confirm basic
stability/robustness. Also ran locally fabric tests locally 200x without
failures
- [x] BH Nightly:
https://github.com/tenstorrent/tt-metal/actions/runs/16243050111
- [x] T3K Unit Tests:
https://github.com/tenstorrent/tt-metal/actions/runs/16206003333
- [x] Metal microbenchmarks:
https://github.com/tenstorrent/tt-metal/actions/runs/16209524287
- [X] New/Existing tests provide coverage for changes