ci: fix permissions of check-stale workflow (#508)
### Ticket
None

### Problem description
We have some stale PRs that we'd like to automatically label as stale.
This workflow has been added some time ago, but after moving to GH
enterprise, it stopped working as expected since it was missing
permissions.

### What's changed
- Added missing permissions to the check-stale workflow.
- Increased number of days before issue becomes stale to 700, since I
wasn't sure about the number already in place (100 days)

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update