[skip ci] Shift Left SFPI Unit Tests (#25255)
### Ticket
NA

### Problem description
These tests spin up a machine in every run of all post commit, to run
for 5 seconds. When you factor in the setup and tear down of the
infrastructure, this gives us at best 5% utilization of the machine
resource. Also, these tests living in all post commit, means that they
are opt in. If we move them to PR-Gate, we guarantee they will never
break.

### What's changed
Bring sfpi unit tests into PR Gate.

### Checklist
- No need to run anything, PR Gate runs automatically on PR.