Check for linked transactions in noc sanitizer (#25307)
### Ticket
#24453

### Problem description
If NOC_CMD_VC_LINKED is set on any command buffer, the next transaction
must be an mcast on the same VC. If not, the NOC will completely fail to
initiate any new transactions.

### What's changed
A new feature is introduced as part of the NOC sanitizer which is
enabled when Watcher is enabled and
TT_METAL_WATCHER_ENABLE_NOC_SANITIZE_LINKED_TRANSACTION is set. It will
flag a NOC sanitization error if a unicast transaction is made when a
linked transaction is still open.

Function `debug_sanitize_post_noc_addr_and_hang` is uninlined. This
error-path function does not need to be inlined, and when compiling some
kernels with `-O3`, it may be inlined, leading to text section overflow.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16357900403/job/46277489759)
CI passes