fix: make tests compile again after change in metal-specific headers (#516)
### Ticket
None

### Problem description

After https://github.com/tenstorrent/tt-metal/pull/25110 was merged,
tests on our main branch stopped working because an additional header
was introduced. This change adds the missing header and reduces
complexity, but for now, the instruction buffer pointer address is
hardcoded until we determine the proper way to define it.

### What's changed
- Added additional header for download
- Hardcoded instruction buffer address
- Removed shim for fetching instruction pointer

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update