Added Dynamic Tiled Implementation For Llama Vision Model (#25117)
### Ticket
#25111

### Problem description
In the current implementation of the vision model, the inputs passed to
the encoder are always of resolution 1120x1120 (max resolution) with the
rest of the image being 0 padded. This is done to enable the graph
compile for prefill stage. A downside of this is that it increases the
prefill time to around 1300ms for all images.

### What's changed
As part of llama's image preprocessing each image is divided into chunks
of 560x560. Before inputs are passed to the encoder we only use the
slice that contains the desired number of chunks. We compile three
different traces for vision prefill for 1, 2 and 4 chunks respectively.
This reduces prefill time for smaller images drastically.

## Results

TTFT for 512x512 image in tt-inference-server has reduced from 1400ms to
500ms and prefill tok/s in `simple_vision_demo.py` increased from 13
tok/s to 24tok/s